Refactor, Feat: Sliding 애니메이션 부분 적용
Sliding 화면전환 방식이 앞으로 여러 위젯에서 쓰일 것 같아 utils 로 위치를 변경하였고 UserPage에서 PostViewPage로 넘어가는 과정에 부분적으로 적용함. 디자이너와 상의 후 전체 적용 또는 원래대로 돌려놓을 예정.
